// src/shared/api/generated/models/unlinkEndUserCaseMessageDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */

export interface UnlinkEndUserCaseMessageDto {
  /** @minimum 1 */
  expectedVersion: number;
  idempotencyKey: string;
  /**
   * @minLength 1
   * @maxLength 2000
   */
  reason: string;
}